+++ src/main/java/com/glencoesoftware/omero/zarr/OmeroAmazonS3ClientFactory.java
@@ -15,7 +15,7 @@
  * with this program; if not, write to the Free Software Foundation, Inc.,
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
-package com.glencoesoftware.omero.ms.core;
+package com.glencoesoftware.omero.zarr;
 
 import java.net.URI;
 import java.util.HashMap;
