+++ src/main/resources/blitz/blitz-ZarrPixelBuffer.xml
@@ -8,7 +8,7 @@
 
   <alias name="${omero.pixeldata.pixels_service:ZarrPixelsService}" alias="/OMERO/Pixels"/>
 
-  <bean name="ZarrPixelsService" class="com.glencoesoftware.omero.ms.core.PixelsService"
+  <bean name="ZarrPixelsService" class="com.glencoesoftware.omero.zarr.ZarrPixelsService"
         parent="filesystem">
     <!-- note constructor-args from parent bean -->
     <constructor-arg value="${omero.pixeldata.memoizer.dir}"/>
